The great fixing - part 1: const correctness and comments

+++ src/score.cpp
@@ -926,7 +926,7 @@ void ScoreDisplayHighScores(  // Returns nothing.
          else
          {
             // Really bad elapsed time.
-            lPlayerScore   = LONG_MAX;
+            lPlayerScore   = INT32_MAX;
          }
 
          vtScoringUnit   = Time;
@@ -942,7 +942,7 @@ void ScoreDisplayHighScores(  // Returns nothing.
          else
          {
             // Really bad elapsed time.
-            lPlayerScore   = LONG_MAX;
+            lPlayerScore   = INT32_MAX;
          }
 
          vtScoringUnit   = Time;
@@ -958,7 +958,7 @@ void ScoreDisplayHighScores(  // Returns nothing.
          else
          {
             // Really bad time.
-            lPlayerScore   = LONG_MIN;
+            lPlayerScore   = INT32_MIN;
          }
 
          vtScoringUnit   = Time;
@@ -974,7 +974,7 @@ void ScoreDisplayHighScores(  // Returns nothing.
          else
          {
             // Really bad time.
-            lPlayerScore   = LONG_MIN;
+            lPlayerScore   = INT32_MIN;
          }
 
          vtScoringUnit   = Time;
@@ -1105,7 +1105,7 @@ void ScoreDisplayHighScores(  // Returns nothing.
             else
             {
                astrTempNames[sIndex][0]   = '\0';
-               alTempScores[sIndex]         = LONG_MIN + 1;
+               alTempScores[sIndex]         = INT32_MIN + 1;
             }
          }
 
@@ -1123,7 +1123,7 @@ void ScoreDisplayHighScores(  // Returns nothing.
          for (sDstIndex = 0; sDstIndex < MAX_HIGH_SCORES; sDstIndex++)
          {
             sHighestScoreIndex   = -1;
-            lHighestScore         = LONG_MIN;
+            lHighestScore         = INT32_MIN;
 
             // Find the highest score of the ones not yet copied.
             for (sSrcIndex = 0; sSrcIndex < MAX_HIGH_SCORES; sSrcIndex++)
@@ -1334,7 +1334,7 @@ void ScoreDisplayHighScores(  // Returns nothing.
 //            rspSetMouseCursorShowLevel(1);
 
             // Make sure there's no timeout on while player is adding their name.
-            ms_lScoreMaxTimeOut   = LONG_MAX;
+            ms_lScoreMaxTimeOut   = INT32_MAX;
 
             // If we want a high score from this player . . .
             if (sPlayersScorePosition >= 0 && pRealm->m_flags.bMultiplayer == false)
@@ -1368,7 +1368,7 @@ void ScoreDisplayHighScores(  // Returns nothing.
             }
             else
             {
-               ms_lScoreMaxTimeOut   = LONG_MAX;
+               ms_lScoreMaxTimeOut   = INT32_MAX;
             }
 
             // Set the focus to the listbox's vertical scrollbar so that the arrows will work.
